+++ src/pages/TeamSettings/TeamSettings.vue
@@ -1,8 +1,10 @@
 <script>
 import { mapMutations, mapGetters } from 'vuex'
-
+import UpgradeBadge from '@/components/UpgradeBadge'
 export default {
-  components: {},
+  components: {
+    UpgradeBadge
+  },
   data() {
     return {
       // Delete team
@@ -27,7 +29,8 @@ export default {
   },
   computed: {
     ...mapGetters('api', ['isCloud']),
-    ...mapGetters('tenant', ['tenant', 'role'])
+    ...mapGetters('tenant', ['tenant', 'role']),
+    ...mapGetters('license', ['hasPermission'])
   },
   watch: {
     tenant() {
@@ -177,6 +180,29 @@ export default {
           </v-list-item-action>
           <v-list-item-content>
             <v-list-item-title>Members</v-list-item-title>
+          </v-list-item-content>
+        </v-list-item>
+
+        <v-list-item
+          :disabled="!isCloud || !hasPermission('feature', 'custom-role')"
+          :to="{ name: 'roles', params: { tenant: tenant.slug } }"
+          ripple
+          exact
+        >
+          <v-list-item-action>
+            <v-icon>face</v-icon>
+          </v-list-item-action>
+          <v-list-item-content>
+            <v-list-item-title
+              >Roles
+              <UpgradeBadge
+                v-if="isCloud && !hasPermission('feature', 'custom-role')"
+                inline
+              >
+                <span class="font-weight-medium">Custom Roles</span> are only
+                available on Enterprise plans.
+              </UpgradeBadge>
+            </v-list-item-title>
           </v-list-item-content>
         </v-list-item>